+++ unshackle/commands/dl.py
@@ -1611,7 +1611,7 @@ class dl:
                         if subtitle.codec == Subtitle.Codec.SubStationAlphav4:
                             for line in subtitle.path.read_text("utf8").splitlines():
                                 if line.startswith("Style: "):
-                                    font_names.append(line.removesuffix("Style: ").split(",")[1])
+                                    font_names.append(line.removeprefix("Style: ").split(",")[1].strip())
 
                     font_count, missing_fonts = self.attach_subtitle_fonts(
                         font_names, title, temp_font_files